+++ src/queue/authentication/QueueSharedKeyAuthenticator.ts
@@ -328,7 +328,7 @@ export default class QueueSharedKeyAuthenticator implements IAuthenticator {
           this.getHeaderValueToSign(req, HeaderConstants.CONTENT_LENGTH),
           this.getHeaderValueToSign(req, HeaderConstants.CONTENT_MD5),
           this.getHeaderValueToSign(req, HeaderConstants.CONTENT_TYPE),
-          this.getHeaderValueToSign(req, HeaderConstants.DATE),
+          this.getHeaderValueToSign(req, HeaderConstants.X_MS_DATE) ? "" : this.getHeaderValueToSign(req, HeaderConstants.DATE),
           this.getHeaderValueToSign(req, HeaderConstants.IF_MODIFIED_SINCE),
           this.getHeaderValueToSign(req, HeaderConstants.IF_MATCH),
           this.getHeaderValueToSign(req, HeaderConstants.IF_NONE_MATCH),
@@ -344,7 +344,7 @@ export default class QueueSharedKeyAuthenticator implements IAuthenticator {
           req.getMethod().toUpperCase(),
           this.getHeaderValueToSign(req, HeaderConstants.CONTENT_MD5),
           this.getHeaderValueToSign(req, HeaderConstants.CONTENT_TYPE),
-          this.getHeaderValueToSign(req, HeaderConstants.DATE)
+          this.getHeaderValueToSign(req, HeaderConstants.X_MS_DATE) ? "" : this.getHeaderValueToSign(req, HeaderConstants.DATE)
         ].join("\n") +
         "\n" +
         this.getCanonicalizedHeadersString(req)