+++ src/table/authentication/TableSharedKeyLiteAuthenticator.ts
@@ -53,8 +53,8 @@ export default class TableSharedKeyLiteAuthenticator implements IAuthenticator {
 
     const stringToSign: string =
       [
-        this.getHeaderValueToSign(req, HeaderConstants.DATE) ||
-          this.getHeaderValueToSign(req, HeaderConstants.X_MS_DATE)
+        this.getHeaderValueToSign(req, HeaderConstants.X_MS_DATE) ||
+          this.getHeaderValueToSign(req, HeaderConstants.DATE)
       ].join("\n") +
       "\n" +
       this.getCanonicalizedResourceString(
@@ -108,8 +108,8 @@ export default class TableSharedKeyLiteAuthenticator implements IAuthenticator {
       // JS/.net Track2 SDK will generate stringToSign from IP style Uri with "-secondary" in authenticationPath, so will also compare signature with this kind stringToSignconst stringToSign: string =
       const stringToSign_secondary: string =
       [
-        this.getHeaderValueToSign(req, HeaderConstants.DATE) ||
-          this.getHeaderValueToSign(req, HeaderConstants.X_MS_DATE)
+        this.getHeaderValueToSign(req, HeaderConstants.X_MS_DATE) ||
+          this.getHeaderValueToSign(req, HeaderConstants.DATE)
       ].join("\n") +
       "\n" +
       this.getCanonicalizedResourceString(